python/pancake-swap-example/requirements.txt
Flask==3.1.0
web3==7.9.0
eth-account==0.13.5
mnemonic==0.21
python-dotenv==1.0.1
flasgger==0.9.7.1
